+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
 
+## [3.0.2] - 06-29-2024
+
+### Added
+
+- unit tests
+
+### Changed
+
+- model deletion returns valid JSON on permission error
+
+### Fixed
+
+- multiple bugs for dataset generation, model creation, and cli
+
 ## [3.0.1] - 06-27-2024
 
 ### Changed
@@ -81,6 +95,7 @@ Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 
 - switched to `pathlib` for file path parsing
 
+[3.0.2]: https://github.com/A-Baji/discordAI-modelizer/compare/3.0.1...3.0.2
 [3.0.1]: https://github.com/A-Baji/discordAI-modelizer/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/A-Baji/discordAI-modelizer/compare/2.0.1...3.0.0
 [2.0.1]: https://github.com/A-Baji/discordAI-modelizer/compare/1.2.2...2.0.1